[CI] Allow MSBuild test stages to run in the megapipeline (#8033)
Makes changes required to get the MSBuild tests running in the unified
pipeline.  New parameters have been added to the the MSBuild test stage
templates to allow them to run without marking the build as failed if
test issues occur.  This was done to ensure that release pipelines that
depend on the unified pipeline can run without being blocked by a known
or intermittent test failure.