[Xamarin.Android.Build.Tasks] missing resource error handling (#8066)
Context: https://github.com/mono/SkiaSharp/issues/2444
Context: https://github.com/mono/SkiaSharp/pull/2465
Context: fcd7cf8f4200ccae92e333c920613bfec260973b
Context: dc3ccf28cdbe9f8c0a705400b83c11a85c81a980
Context: 3ec1b159b8125cfaea1d774da94b323f08e2432b

In Classic Xamarin.Android, `@(AndroidResource)` files were embedded
into assemblies as `@(EmbeddedResource)` entries.  While convenient,
this slowed down build times (as every assembly needed to be checked
to see if it had embedded resources), and this functionality was
removed for .NET 6 in fcd7cf8f.  The replacement was to use `.aar`
files, typically located in the same directory as the assembly.

Unfortunately, it was easy to incorrectly migrate from the old system
to the new system, which was the case for the 
[SkiaSharp.Views (v2.88.3) NuGet Package][0], which *uses* an
[`@attr/ignorePixelScaling`][1] resource, but did not package it.

In .NET 6 and 7, this was "fine" *so long as* that resource value was
never used.  If it *was* used, then the results were undefined: a
"garbage" resource id value would be used, which would either be
invalid or refer to a *different* resource altogether!

.NET 8 introduces the Resource Designer Assembly (dc3ccf28), which
involves *rewriting non-.NET 8 assemblies* to use the Resource
Designer Assembly, which in turn requires that all resources exist.

Which brings us to mono/SkiaSharp#2444: because the `SkiaSharp.Views`
NuGet package references but did not redistribute the
`@attr/ignorePixelScaling` resource, the assembly rewriter would
*skip over* the missing resource name, resulting in an invalid
assembly.  This in turn would result in a *runtime crash*:

	System.BadImageFormatException: 'Invalid field token 0x040000a1'

We decided that instead of ignoring missing resource fields, we
should error out.

Update `FixLegacyResourceDesignerStep` to report an error if we
cannot find the required Resource Designer Assembly property which
matches a "legacy" field.

In Debug builds, this will result in an XA8000 error:

	error XA8000: Could not find Android Resource '@attr/ignorePixelScaling'.
	Please update @(AndroidResource) to add the missing resource.

For Release builds, this instead results in an IL8000 error:

	error IL8000: Could not find Android Resource '@attr/ignorePixelScaling'.
	Please update @(AndroidResource) to add the missing resource.

Note that this only moves a runtime error to a compile-time error.
The fix for mono/SkiaSharp#2444 is mono/SkiaSharp#2465, which adds
`SkiaSharp.Views.Android.aar` to the `SkiaSharp.Views` NuGet package.
This should be fixed in the (forthcoming) SkaSharp.Views 2.88.4.
[A *workaround* for mono/SkiaSharp#2444][3] is to define the
`@attr/ignorePixelScaling` resource within your app project.

Additionally, we found another problem while investigating
mono/SkiaSharp#2444 around resource name case sensitivity.  Not all
Android Resource names need to be lowercase; generally only those
that are based on *filenames* need to be lowercase, e.g.
`@layout/main`.  Resources which are defined within XML containers,
such as [string resources][2], may contain uppercase letters.

While looking at the output for the test app, it turns out the final
assembly does NOT include the expected `anim` based resources.
Looking at the IL we found that the properties were NOT being fixed
up to use the correct casing, e.g. we have `enterfromright` instead
of `EnterFromRight`.  Further investigation tracked the issue down to
the `RTxtParser`; This class uses the `case_map.txt` file for lookup.
The `case_map.txt` file contains entries such as:

	animation/enterfromleft;animation/EnterFromLeft
	animation/enterfromright;animation/EnterFromRight

these allow us to map the "android" lower cased items to the C# cased
items.   The problem was we were not mapping `anim` to `animation`.
There is a method called `ResourceParser.GetNestedTypeName()` which
is used to map the android resource type to the managed resource type.
For example it maps `anim` to `Animation` and `bool` to `Boolean`.
This was not being called when we tried to map the resource.  As a
result we never found the correct cased entry.  This would result in
the final Resource Designer Assembly containing the lower cased
"android" naming of the property and not the correctly cased one
which the C#/IL would be using; the fields were never fixed up. 

The fix is to make sure we call `ResourceParser.GetNestedTypeName()`
when calculating the final property name in the `RTxtParser`.

[0]: https://www.nuget.org/packages/SkiaSharp/2.88.3
[1]: https://github.com/mono/SkiaSharp/blob/99c2437b1055dc6eb8ee25e90e5ad4afa695aa89/source/SkiaSharp.Views/SkiaSharp.Views.Android/Resources/values/attrs.xml#L4
[2]: https://developer.android.com/guide/topics/resources/string-resource
[3]: https://github.com/mono/SkiaSharp/issues/2444#issuecomment-1564090245