[build] stop setting `$DOTNET_gcServer` (#8085)
Context: https://github.com/dotnet/linker/issues/3012
Context: https://github.com/dotnet/runtime/issues/86807

In 34986b9a, we exported `DOTNET_gcServer=0` so that the Server GC is disabled, in an
effort to prevent an `AccessViolationException`.

It appears that this might be *fixed* in .NET 8. Let's remove the workaround, so we
can either close the issue or get further investigation on the issue.