[xaprepare] Add support for Debian trixie release
Context: https://bits.debian.org/2023/06/bookworm-released.html
Context: https://siduction.org/

Debian 12, aka bookworm, has just been released and that means work on
the next release, codenamed "trixie", has begun.  The code name appeared
in `/etc/debian_version` and `/etc/os-release` files and `xaprepare`
needs to be updated in order to properly detect trixie.

While this change won't be necessary for users of Debian "stable" or any
distributions based on it, it is necessary for rolling Debian "unstable"
distributions like `siduction` (which is what I use :D)

Debian "unstable" distribution (which eventually becomes the "testing"
distribution, to become "stable" on release) does not use version
numbers in the abovementioned files, just the codename.  This is the
reason we need to fake the OS version number for Debian.