[marshal methods] Add support for wrapping the `char` type
`System.Char` is an unblittable type and, as such, it has to be treated
specially by the marshal methods wrapper generator to properly cast it
to/from a matching blittable type.  In this case the latter is `ushort`,
since both on the Java and Managed side, `char` is a 16-bit entity.