[AndroidDependenciesTests] Test both manifest types
Updates the InstallAndroidDependenciesTest test to run against the self
hosted Xamarin manifest, as well as Googles v2 manifest.

For the default case, we shouldn't override any package versions.  The
dependency versions we calculate should be present in, and installable
from, our manifest.

For the Google manifest case, we should override the platform-tools
version.  Their manifest only ever contains one platform-tools version,
and it is updated somewhat regularly causing the test to fail every time
it changes.