Clean up DotNetIgnore Unit Tests (#8163)
We have recently begun to remove suport for Xamarin.Android Classic.
Many of the Unit Tests labeled with the "DotNetIgnore" Category were flagged
as such because they did not work under .Net Android. This commit goes through
and removes tests which are not relevant in the .Net Android context. It also
modifies some of these tests which are relevant to work under .Net Android.