[ci] Disable .NET jobs by default
Disables all of the jobs that are unique to .NET Android NuGet package
creation, testing, and signing, as we we do not plan to release new
nuget packages from the `d17-5` branch.

This behavior is controlled by a new pipeline parameter named
`createSignNupkgs` which can be set to `true` if it's ever needed.