[profiled-aot] update profile for .NET 8.0.100-rc.1.23404.2 (#8250)
Using the latest bump from e9b19e48, I updated `dotnet.aotprofile`.

An average of 10 runs of a `dotnet new android` app on a Pixel 5:

    Before:
    Average(ms): 199.2
    Std Err(ms): 2.40277617212535
    Std Dev(ms): 7.59824541149688
    After:
    Average(ms): 185.1
    Std Err(ms): 2.54492959955018
    Std Dev(ms): 8.04777401935881

It seems be about 15ms better with the new profile. It looks like we
last updated this in .NET 8 Preview 5, but it was down to 166.3ms back
then.

I will make a similar change in .NET MAUI next.