[api-merge] We only need to write API-34 api.xml. (#8248)
In Classic, we needed to build a `Mono.Android.dll` for every Android API level we supported.  In .NET we only build the latest API level(s).  Thus we do not need to output an `api.xml` for API levels we aren't building.

Save a few build seconds and several hundred megabytes of disk space by not writing unneeded `api.xml` files.