Update dotnet-local.cmd
Take…3?

The problem: running `dotnet-local.cmd build path\to\app.csproj` would error out on CI with:

    \NSIS\ was unexpected at this time.
    ##[error]Process completed with exit code 255.

Running locally was no better:

    \Microsoft was unexpected at this time.

like, what?

Eventually, after a bit of experimentation and "man page" reading, we see the `SET /?` output:

```
Finally, support for delayed environment variable expansion has been
added.  This support is always disabled by default, but may be
enabled/disabled via the /V command line switch to CMD.EXE.  See CMD /?

Delayed environment variable expansion is useful for getting around
the limitations of the current expansion which happens when a line
of text is read, not when it is executed.  The following example
demonstrates the problem with immediate variable expansion:

    set VAR=before
    if "%VAR%" == "before" (
        set VAR=after
        if "%VAR%" == "after" @echo If you see this, it worked
    )

would never display the message, since the %VAR% in BOTH IF statements
is substituted when the first IF statement is read, since it logically
includes the body of the IF, which is a compound statement.  So the
IF inside the compound statement is really comparing "before" with
"after" which will never be equal.  Similarly, the following example
will not work as expected:

    set LIST=
    for %i in (*) do set LIST=%LIST% %i
    echo %LIST%

in that it will NOT build up a list of files in the current directory,
but instead will just set the LIST variable to the last file found.
Again, this is because the %LIST% is expanded just once when the
FOR statement is read, and at that time the LIST variable is empty.
So the actual FOR loop we are executing is:

    for %i in (*) do set LIST= %i

which just keeps setting LIST to the last file found.

Delayed environment variable expansion allows you to use a different
character (the exclamation mark) to expand environment variables at
execution time. …
```

:thinking: 

As a quick experiment, consider this fragment:

```
IF EXIST "%ROOT%\bin\Release\dotnet\dotnet.exe" (
    SET XA_DOTNET_ROOT=%ROOT%\bin\Release\dotnet
    echo # jonp: XA_DOTNET_ROOT=%XA_DOTNET_ROOT%
)
```

The result is that it would print:

```
# jonp: XA_DOTNET_ROOT=
```

i.e. *the empty string*!

My hypothesis is that because environment expansion was "immediate" (not "delayed"), attempting to do `SET PATH=%XA_DOTNET_ROOT%;%PATH%` would result in "weird recursion", and thus all the errors.

The fix?  Avoid immediate environment variable expansion, by LIMITING what is done in statement for `IF CONDITION STATEMENT`.  Instead of executing, *only* set env vars:

```
IF EXIST "%ROOT%\bin\Release\dotnet\dotnet.exe" (
    SET XA_DOTNET_ROOT=%ROOT%\bin\Release\dotnet
    SET XA_CONFIG=Release
) …
```

Then when we later try `SET PATH=%XA_DOTNET_ROOT%;%PATH%`, things are "saner", and -- at least locally -- we shouldn't get those "bizarre" `… was unexpected at this time` messages.

Finally, use `SETLOCAL` so that our updates to `%PATH%` don't "overwrite" the calling environment.