[build] set file extension for common `ToolExe` values
Running the `-t:Install` target with a local xamarin-android build on
Windows, I got the error:

    AndroidAdb
    Errors
        Xamarin.Android.Common.targets(2641,3): error MSB6004:
        The specified task executable location "D:\android-toolchain\sdk\platform-tools\adb" is invalid.

This appears to only happen with a local of xamarin/xamarin-android
because it specifies `AdbToolExe=adb` in `Configuration.props`.

It appears that most of our tasks just use the pattern:

    protected override string GenerateFullPathToTool ()
    {
        return Path.Combine (ToolPath, ToolExe);
    }

Where if `ToolExe` was completely blank, it would fall back to the value:

    protected override string ToolName => OS.IsWindows ? "adb.exe" : "adb";

Searching the web:

https://grep.app/search?current=2&q=ToolExe&words=true&filter[lang][0]=C%23&filter[lang][1]=XML

It seems most other tasks use `ToolExe` as-is, if it is passed in. Where
most of the time the default value is blank.

So, I think the simple fix here, is to specify the file extension in
`Configuration.props` for any of our `$(FooToolExe)` values.

I also removed `$(AndroidToolExe)` because it is unused and may not
event exist anymore in the Android SDK.