[Xamarin.Android,Build.Tasks] Cannot merge Android Manifest elements from class library
Fixes https://github.com/xamarin/xamarin-android/issues/8262

So if a user wants to distribute an AndroidManifest.xml file along with
a NuGet the file is currently NOT included in the library .aar file.
So lets include it if one is present. This will allow NuGet authors
to set default permissions if need be by simply adding the file to the
project directory.

This also fixes an issue in GenerateLibraryResources where if a manifest
file does NOT have a package name.