[Xamarin.Android.Build.Tasks] Fix APT2264 error on Windows.
We see users reporting the following error on Windows machines.

```
APT2000 The system cannot find the file specified. (2): foo.xml
```

This is odd because what it SHOULD be reporting is an APT2264
error. This provides additional feedback to check for Long File Paths.
As a result users do not get the additional info.
It turns out we were trying to match
`The system cannot find the file specified. (2).`
with `The system cannot find the file specified. (2):`.
On Windows the error message contains a `:` rather than `.` this
stops the code matching and results in the APT2000 error.
Lets fix that.