[Xamarin.Android.Build.Tasks] Fix APT2264 error on Windows. (#8289)
We see users reporting the following error on Windows machines.

	APT2000 The system cannot find the file specified. (2): foo.xml

This is odd because what it *should* be reporting an APT2264 error.
This provides additional feedback to check for Long File Paths.  As a
result users do not get the additional info.

It turns out that `aapt2` generates subtly different error messages
between Linux/macOS and Windows, replacing a `.` with `:`.

  * Linux/macOS:

        The system cannot find the file specified. (2).

  * Windows
        
        The system cannot find the file specified. (2):

Because our APT2264-checking code included the trailing `.` and
Windows instead had a trailing `:`, we would *never* return APT2264
on Windows.

Lets fix that by removing the trailing `.` from what we look for.