[xaprepare] Fix dotnet-install script size check (#8311)
CI builds have been failing frequently and quietly:

    Downloading dotnet-install...
      Error: Installation of dotnet SDK 8.0.100-rc.2.23422.31 failed.

    Step Xamarin.Android.Prepare.Step_InstallDotNetPreview failed
    System.InvalidOperationException: Step Xamarin.Android.Prepare.Step_InstallDotNetPreview failed

We were returning early when attempts to calculate the size of the
`dotnet-install` script failed, and a cached install script file was not
already on disk.  This has been fixed by ensuring that we still attempt
to download the script if the size calculation fails.

The size check failure has also been fixed by using a get request, and
logging around this prepare step has been improved.