[Mono.Android] Add missing `[Flags]` attribute for generated enums. (#8310)
Fixes: https://github.com/xamarin/xamarin-android/issues/7416

We were hoping we could use the data that Google provides in
`annotations.xml` to create a better process for identifying which
sets of constants should be a `[Flags]` enum.  However, the data is
too dirty to create anything automatically (incorrect data from Google,
constants groups that don't map 1:1 to our enums, etc.).

However, it did provide a large list of potential candidates.  These
candidates were manually compared to the Android documentation and
appropriate ones have been updated to be generated with the `[Flags]`
attribute.