Add a target to run application with logging
We are frequently faced with the need to look at some detailed
information optionally logged by our runtime (as well as by the MonoVM
runtime in dotnet) and in order to obtain the necessary data, we ask our
users to perform a series of steps on command line.

This is a task we can easily wrap in a nice target, so that instead of
having to type 5 lines of commands the user invokes a single command
which will do the rest for them:

    dotnet build -t:RunWithLogging

The above command will set the `debug.mono.log` property to the desired
value, increase logcat buffer, clear the buffer, start the application
waiting for it to be fully started, then pause for 1s and dump the
logcat buffer to a file.

Optionally, the `$(_RunLogVerbose)` property can be set to `true` in
order to log verbose MonoVM data and the `$(_RunLogFilePath)` can be set
to a logcat output file path (defaults to
`$(IntermediateOutputPath)/logcat.txt`)