New direction - to per-assembly DSOs
Some assemblies will still be contained in libxamarin-app.so,
but most of them will end up in their own shared libraries.
Te reason is memory usage (so that we don't have to keep the
compressed assembly data in memory indefinitely).  The per-assembly
shared libraries will not be `dlopen`-ed because that call is too
expensive (half a millisecond on Pixel 6 Pro, 1.6ms on Nexus 5) while
mmap is very cheap, so we will record, at build time, location of the
individual assembly's data in its DSO and use that to read the content
after mmapping it.

This requires some reworking of the current code as the sequence in
which the DSOs are generated needs to be done in three steps:

  * generate the individual assemblies DSO source code
  * compile the above and link the DSOs
  * generate code for inclusion in libxamarin-app.so

This is because, before the last step, we need to know the location
of data in the DSOs, and that data is available only when we have
the .so library and can read it with ELFSharp.