[installers] Let monodroid control which fast dev files are included in the installer (#8297)
Related xamarin/monodroid@a2dbe75.

Currently we control which files are included from the monodroid Fast Deployment in this repo.
This can cause problems when we want to add new files (or remove them). Since we would need
to bump each repo multiple times to get everything to work.

This PR moves that control over to monodroid, so new files can be included in one commit and then
one bump.